+++ packages/utils/src/rundown-utils/rundownUtils.test.ts
@@ -2,7 +2,6 @@ import { OntimeRundown, SupportedEvent } from 'ontime-types';
 
 import { dayInMs } from '../timeConstants.js';
 import { getNextEvent, getPreviousEvent } from './rundownUtils';
-import { calculateDuration } from './rundownUtils.js';
 
 describe('getNextEvent()', () => {
   it('returns the next event of type event', () => {
@@ -69,27 +68,5 @@ describe('getPreviousEvent()', () => {
 
     const previous = getNextEvent(testRundown as OntimeRundown, '1');
     expect(previous).toBe(null);
-  });
-});
-
-describe('calculateDuration()', () => {
-  describe('Given start and end values', () => {
-    it('is the difference between end and start', () => {
-      const duration = calculateDuration(10, 20);
-      expect(duration).toBe(10);
-    });
-  });
-
-  describe('Handles edge cases', () => {
-    it('handles events that go over midnight', () => {
-      const duration = calculateDuration(51, 50);
-      expect(duration).toBe(dayInMs - 1);
-    });
-    it('handles no difference', () => {
-      const duration1 = calculateDuration(0, 0);
-      const duration2 = calculateDuration(dayInMs, dayInMs);
-      expect(duration1).toBe(0);
-      expect(duration2).toBe(0);
-    });
   });
 });
